Create favicon.ico
I added a favicon to the upper corner of the tab, and it's a star.

A favicon (short for "favorites icon") is a small, 16x16 pixel icon displayed in a browser's address bar, tab, or bookmark menu next to a website's name. It helps users quickly identify and locate your website among the many tabs or bookmarks they may have open.